case_studies/ey-olympiad-hackathon-2024/requirements.txt
pandas
beautifulsoup4
rapidocr-onnxruntime
pypdf
langchain_core
langchain_community
langchain_openai
langchain-experimental
langchain-google-genai
langserve
langchainhub
PyMuPDF 
Pillow
pdf2image
pdfminer.six
streamlit

python-dotenv
sse_starlette

azure-storage-blob
azure-identity
azure-search-documents==11.4.0b6
azure-ai-translation-text
azure-ai-textanalytics==5.2.0

urllib3
ipykernel
unstructured